backend/requirements/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
websockets==12.0
kafka-python==2.0.2
boto3==1.34.11
itsdangerous==2.1.2
authlib==1.3.0
httpx==0.26.0
openai==1.6.1